+++ src/PockyBot.NET/Constants/Commands.cs
@@ -4,9 +4,12 @@ namespace PockyBot.NET.Constants
     {
         public const string Peg = "peg";
         public const string Ping = "ping";
+        public const string Help = "help";
+        public const string Welcome = "welcome";
         public const string Status = "status";
         public const string Finish = "finish";
         public const string Reset = "reset";
+        public const string Rotation = "rotation";
         public const string LocationConfig = "locationconfig";
     }
 }