+++ src/PockyBot.NET/Constants/Roles.cs
@@ -1,6 +1,7 @@
 namespace PockyBot.NET.Constants
 {
-    internal static class Roles
+    internal static class
+        Roles
     {
         public const string Unmetered = "UNMETERED";
         public const string Admin = "ADMIN";